Fold ScipSymbol into ScipPrinters
